feat: solucao imagens

+++ src/mundo-veg/Controllers/Produtoscontroller.cs
@@ -13,11 +13,9 @@ namespace mundo_veg.Controllers
     public class ProdutosController : Controller
     {
         private readonly ApplicationDbContext _context;
-        private string _filePath;
 
-        public ProdutosController(ApplicationDbContext context, IWebHostEnvironment env)
+        public ProdutosController(ApplicationDbContext context)
         {
-            _filePath = env.WebRootPath;
             _context = context;
         }
 
@@ -77,18 +75,13 @@ namespace mundo_veg.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Nome,Quantidade,Descricao,Preco,Categoria,Imagem")] Produto produto, IFormFile anexo)
+        public async Task<IActionResult> Create([Bind("Id,Nome,Quantidade,Descricao,Preco,Categoria")] Produto produto)
         {
             var user = HttpContext.User;
             var userId = user.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Sid).Value;
             produto.EstabelecimentoId = int.Parse(userId);
             if (ModelState.IsValid)
             {
-                if (!ValidaImagem(anexo))
-                    return View(produto);
-                var nomeImg = SalvarArquivo(anexo);
-                produto.Imagem = nomeImg;
-
                 _context.Add(produto);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -97,42 +90,6 @@ namespace mundo_veg.Controllers
                 ViewBag.Message = "Erro ao cadastrar produto!";
             }
             return View(produto);
-        }
-        public bool ValidaImagem(IFormFile anexo)
-        {
-            switch (anexo.ContentType)
-            {
-                case "image/jpg":
-                    return true;
-                case "image/jpeg":
-                    return true;
-                case "image/bmp":
-                    return true;
-                case "image/png":
-                    return true;
-                default:
-                    return false;
-                    break;
-
-            }
-        }
-        public string SalvarArquivo(IFormFile anexo)
-        {
-            var nome = Guid.NewGuid().ToString() + anexo.FileName;
-
-            var filePath = _filePath + "\\imagens";
-
-            if (!Directory.Exists(filePath)) ;
-            {
-                Directory.CreateDirectory(filePath);
-            }
-
-            using (var stream = System.IO.File.Create(filePath + "\\" + nome))
-            {
-                anexo.CopyToAsync(stream);
-            }
-            return nome;
-
         }
 
         // GET: Produtos/Edit/5
@@ -156,7 +113,7 @@ namespace mundo_veg.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Nome,Quantidade,Ingredientes,Preco,Categoria,Imagem")] Produto produto, IFormFile anexo)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Nome,Quantidade,Ingredientes,Preco,Categoria")] Produto produto)
         {
             if (id != produto.Id)
             {
@@ -167,13 +124,6 @@ namespace mundo_veg.Controllers
             {
                 try
                 {
-
-                    if (!ValidaImagemEdit(anexo))
-                        return View(produto);
-                    var nomeImg = SalvarArquivo(anexo);
-                    produto.Imagem = nomeImg;
-                
-
                     _context.Update(produto);
                     await _context.SaveChangesAsync();
                 }
@@ -192,42 +142,7 @@ namespace mundo_veg.Controllers
             }
             return View(produto);
         }
-        public bool ValidaImagemEdit(IFormFile anexo)
-        {
-            switch (anexo.ContentType)
-            {
-                case "image/jpg":
-                    return true;
-                case "image/jpeg":
-                    return true;
-                case "image/bmp":
-                    return true;
-                case "image/png":
-                    return true;
-                default:
-                    return false;
-                    break;
 
-            }
-        }
-        public string SalvarArquivoEdit(IFormFile anexo)
-        {
-            var nome = Guid.NewGuid().ToString() + anexo.FileName;
-
-            var filePath = _filePath + "\\imagens";
-
-            if (!Directory.Exists(filePath)) ;
-            {
-                Directory.CreateDirectory(filePath);
-            }
-
-            using (var stream = System.IO.File.Create(filePath + "\\" + nome))
-            {
-                anexo.CopyToAsync(stream);
-            }
-            return nome;
-
-        }
 
         // GET: Produtos/Delete/5
         public async Task<IActionResult> Delete(int? id)
@@ -252,15 +167,18 @@ namespace mundo_veg.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+
+            if (_context.Produtos == null)
+            {
+                return Problem("Entity set 'ApplicationDbContext.Produtos'  is null.");
+            }
             var produto = await _context.Produtos.FindAsync(id);
-            string filePathName = _filePath + "\\imagens\\" + produto.Imagem;
 
+            if (produto != null)
+            {
+                _context.Produtos.Remove(produto);
+            }
 
-            if (System.IO.File.Exists(filePathName))
-                System.IO.File.Delete(filePathName);
-            _context.Produtos.Remove(produto);
-            
-            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
